판매가 수정

+++ team_mypet/src/main/java/com/jeonju/mypet/controller/SellerController.java
@@ -50,8 +50,8 @@ public class SellerController {
 			Model model, HttpServletRequest request) {
 		
 		HttpSession session = request.getSession();
-//		int midx = (int) session.getAttribute("midx");
-		long midx = (long) session.getAttribute("midx");
+		int midx = (int) session.getAttribute("midx");
+//		long midx = (long) session.getAttribute("midx");
 		
 		String member_id= Integer.toString((int) midx);
 	//	String member_id= Integer.toString(midx);
